+++ turbot/provider.go
@@ -49,6 +49,7 @@ func Provider() terraform.ResourceProvider {
 			"turbot_grant":                   resourceTurbotGrant(),
 			"turbot_grant_activation":        resourceTurbotGrantActivation(),
 			"turbot_turbot_directory":        resourceTurbotTurbotDirectory(),
+			"turbot_file":                    resourceTurbotFile(),
 		},
 		DataSourcesMap: map[string]*schema.Resource{
 			"turbot_policy_value": dataSourceTurbotPolicyValue(),